[Refactor] move `SmartStrings` into `Strings` namespace
