+++ src/base/common/utilities/string.ts
@@ -3,6 +3,7 @@ import { Result } from "src/base/common/result";
 import { Iterable } from "src/base/common/utilities/iterable";
 import { compareSubstringIgnoreCase } from "src/base/common/files/glob";
 import { CompareOrder, isObject } from "src/base/common/utilities/type";
+import { OS_CASE_SENSITIVE } from "src/base/common/platform";
 
 /**
  * @namespace Strings A collection of functions that relates to string types.
@@ -480,6 +481,34 @@ export namespace Strings {
             tagName: tagName,
             attributes: anyAttributes ? attributes : null
         };
+    }
+
+    export namespace Smart {
+    
+        /**
+         * @description If case sensitive, return the same string, otherwise
+         * a lower case version of the string returned.
+         */
+        export function adjust(str: string): string {
+            if (OS_CASE_SENSITIVE) {
+                return str;
+            }
+            return str.toLowerCase();
+        }
+    
+        export function equals(a: string, b: string): boolean {
+            if (OS_CASE_SENSITIVE) {
+                return a === b;
+            }
+            return Strings.IgnoreCase.equals(a, b);
+        }
+    
+        export function startsWith(str: string, candidate: string): boolean {
+            if (OS_CASE_SENSITIVE) {
+                return str.startsWith(candidate);
+            }
+            return Strings.IgnoreCase.startsWith(str, candidate);
+        }
     }
 }
