+++ test/base/common/utilities/string.test.ts
@@ -1,4 +1,5 @@
 import * as assert from 'assert';
+import { IS_LINUX } from 'src/base/common/platform';
 import { Strings } from 'src/base/common/utilities/string';
 
 suite('Strings-test', function () {
@@ -561,6 +562,62 @@ suite('Strings-test', function () {
                 type: 'self-closing',
                 tagName: 'br',
                 attributes: null
+            });
+        });
+    });
+
+    suite('Smart', function () {
+        
+        suite('#adjust()', function () {
+            test('CaseSensitive', function () {
+                if (!IS_LINUX) {
+                    this.skip();
+                }
+                assert.strictEqual(Strings.Smart.adjust('HeLLo'), 'HeLLo');
+            });
+            
+            test('CaseIgnore', function () {
+                if (IS_LINUX) {
+                    this.skip();
+                }
+                assert.strictEqual(Strings.Smart.adjust('HeLLo'), 'hello');
+                assert.strictEqual(Strings.Smart.adjust('hello'), 'hello');
+            });
+        });
+    
+        suite('#equals()', function () {
+            test('CaseSensitive', function () {
+                if (!IS_LINUX) {
+                    this.skip();
+                }
+                assert.strictEqual(Strings.Smart.equals('hello', 'hello'), true);
+                assert.strictEqual(Strings.Smart.equals('hello', 'HELLO'), false);
+            });
+    
+            test('CaseIgnore', function () {
+                if (IS_LINUX) {
+                    this.skip();
+                }
+                assert.strictEqual(Strings.Smart.equals('hello', 'hello'), true);
+                assert.strictEqual(Strings.Smart.equals('hello', 'HELLO'), true);
+            });
+        });
+    
+        suite('#startsWith()', function () {
+            test('CaseSensitive', function () {
+                if (!IS_LINUX) {
+                    this.skip();
+                }
+                assert.strictEqual(Strings.Smart.startsWith('HELLO world', 'Hello'), false);
+                assert.strictEqual(Strings.Smart.startsWith('Hello world', 'Hello'), true);
+            });
+    
+            test('CaseIgnore', function () {
+                if (IS_LINUX) {
+                    this.skip();
+                }
+                assert.strictEqual(Strings.Smart.startsWith('HELLO world', 'hello'), true);
+                assert.strictEqual(Strings.Smart.startsWith('hello world', 'hello'), true);
             });
         });
     });
